# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Genesis Lists API
  version: 1.0.0
  description: |
    REST API for Genesis Lists MVP. Session cookie authentication.
    Spec is the source of truth — see docs/04-api.md.
servers:
  - url: /
paths:
...
  /api/auth/register:
    post:
      operationId: register
      summary: Register a new user and create a session
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
            example:
              username: alice
              password: secretpass
      responses:
        '201':
          description: Created and logged in
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Session cookie (genesis_session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/auth/login:
    post:
      operationId: login
      summary: Log in and create a session
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '200':
          description: Logged in
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/logout:
    post:
      operationId: logout
      summary: Destroy the current session
      tags: [Auth]
      security:
        - sessionCookie: []
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/me:
    get:
      operationId: getMe
      summary: Current authenticated user
      tags: [Auth]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/change-password:
    post:
      operationId: changePassword
      summary: Change the current user's password
      description: |
        Updates the password hash. Other sessions for this user are deleted;
        the session that made this request stays signed in.
      tags: [Auth]
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password updated
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: genesis_session
...
  schemas:
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_ERROR
            message:
              type: string
    AuthCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9_-]+$'
        password:
          type: string
          minLength: 8
          maxLength: 128
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
          maxLength: 128
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request body
    UnauthorizedError:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
...
    ConflictError:
      description: Conflict (e.g. username taken)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              code: CONFLICT
              message: Username already taken
    ForbiddenError:
      description: Authenticated or policy refusal (for example registration closed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              code: FORBIDDEN
              message: Registration is closed